Consolidate namespaces for serialization tests
