+++ test/NonSilo.Tests/Serialization/CustomSerializerTests.cs
@@ -1,14 +1,10 @@
-﻿using System;
-using Orleans.CodeGeneration;
-using Orleans.Runtime;
-using Orleans.Runtime.Configuration;
-using Orleans.Serialization;
+﻿using Orleans.Serialization;
 using TestExtensions;
 using UnitTests.GrainInterfaces;
 using UnitTests.Grains;
 using Xunit;
 
-namespace NonSilo.Tests.UnitTests.SerializerTests
+namespace UnitTests.Serialization
 {
     [Collection(TestEnvironmentFixture.DefaultCollection)]
     public class CustomSerializerTests
